+++ Project1/p1d.cpp
@@ -46,7 +46,7 @@ int main(int argc, char *argv[]){
 
                 //loop to update
                 for(int i = 0; i < n+2; i++) {
-                        x[i] = double(i)*h;
+                        x[i] = i*h;
                         fprime[i] = f(x[i],h);
                 }
